fix: moving DendronTreeViewV1 to utilize new engine events
