+++ packages/plugin-core/src/WorkspaceWatcher.ts
@@ -405,7 +405,9 @@ export class WorkspaceWatcher {
       Sentry.captureException(error);
       throw error;
     } finally {
-      FileWatcher.refreshTree();
+      // TODO: Remove this call altogether once backlinks updates via
+      // EngineEvents
+      FileWatcher.refreshBacklinks();
     }
   }
