+++ packages/plugin-core/src/views/DendronTreeViewV2.ts
@@ -20,6 +20,11 @@ import { VSCodeUtils } from "../vsCodeUtils";
 import { WSUtils } from "../WSUtils";
 import { WebViewUtils } from "./utils";
 
+/**
+ * Class managing the webview version of the Dendron tree view (currently called
+ * TreeViewV2) - this is the side panel UI that gives the webview/react/antd
+ * based tree view of the Dendron note hierarchy
+ */
 export class DendronTreeViewV2 implements vscode.WebviewViewProvider {
   public static readonly viewType = DendronTreeViewKey.TREE_VIEW_V2;
