+++ packages/plugin-core/src/views/NativeTreeView.ts
@@ -1,188 +1,108 @@
 import {
-  DendronError,
-  DNodeUtils,
+  DendronTreeViewKey,
   NoteProps,
-  NotePropsDict,
   NoteUtils,
+  VaultUtils,
 } from "@dendronhq/common-all";
+import { WorkspaceUtils } from "@dendronhq/engine-server";
 import _ from "lodash";
-import vscode, { ProviderResult, ThemeIcon } from "vscode";
-import { ICONS } from "../constants";
+import path from "path";
+import { Disposable, TextEditor, TreeView, window } from "vscode";
 import { ExtensionProvider } from "../ExtensionProvider";
-import { Logger } from "../logger";
 import { EngineEvents } from "../services/EngineEventService";
-import { TreeNote } from "./DendronTreeView";
+import { EngineNoteProvider } from "./EngineNoteProvider";
 
-export class EngineNoteProviderV2
-  implements vscode.TreeDataProvider<NoteProps>
-{
-  private _onDidChangeTreeDataEmitter: vscode.EventEmitter<
-    NoteProps | undefined | void
-  >;
-  readonly onDidChangeTreeData: vscode.Event<NoteProps | undefined | void>;
-  private _tree: { [key: string]: TreeNote } = {};
+/**
+ * Class managing the vscode native version of the Dendron tree view - this is
+ * the side panel UI that gives a tree view of the Dendron note hierarchy
+ */
+export class NativeTreeView implements Disposable {
+  private treeView: TreeView<NoteProps> | undefined;
+  private _handler: Disposable | undefined;
+  private _events: EngineEvents;
 
-  private _engineEvents;
-  // private sort(notes: TreeNote[]): TreeNote[] {
-  //   return _.sortBy(notes, "label");
-  // }
-
-  constructor(engineEvents: EngineEvents) {
-    this._onDidChangeTreeDataEmitter = new vscode.EventEmitter<
-      NoteProps | undefined | void
-    >();
-
-    this.onDidChangeTreeData = this._onDidChangeTreeDataEmitter.event;
-
-    this._engineEvents = engineEvents;
-
-    this.setupSubscriptions();
+  constructor(events: EngineEvents) {
+    this._events = events;
   }
-
-  private setupSubscriptions(): void {
-    this._engineEvents.onNoteCreated((noteProps) => {
-      //TODO add noteprops payload to optimize
-      this._onDidChangeTreeDataEmitter.fire();
-    });
-
-    this._engineEvents.onNoteDeleted((noteProps) => {
-      //TODO add noteprops payload to optimize
-      this._onDidChangeTreeDataEmitter.fire();
-    });
-  }
-
-  private sortChildren(children: string[], noteDict: NotePropsDict) {
-    return _.sortBy(children, (id) => noteDict[id].title);
-  }
-
-  getTreeItem(noteProps: NoteProps): vscode.TreeItem {
-    return this._tree[noteProps.id];
-  }
-
-  getChildren(noteProps?: NoteProps): ProviderResult<NoteProps[]> {
-    const ctx = "TreeView:getChildren";
-    Logger.debug({ ctx, id: noteProps });
-    const { engine } = ExtensionProvider.getDWorkspace();
-    const roots = _.filter(_.values(engine.notes), DNodeUtils.isRoot);
-    if (!roots) {
-      vscode.window.showInformationMessage("No notes found");
-      return Promise.resolve([]);
+  dispose() {
+    if (this._handler) {
+      this._handler.dispose();
+      this._handler = undefined;
     }
-    if (noteProps) {
-      const childrenIds = noteProps.children;
 
-      const childrenNoteProps = childrenIds.map((id) => {
-        return engine.notes[id];
+    if (this.treeView) {
+      this.treeView.dispose();
+      this.treeView = undefined;
+    }
+  }
+
+  /**
+   * Creates the Tree View and shows it in the UI (registers with vscode.window)
+   */
+  async show() {
+    const treeDataProvider = new EngineNoteProvider(this._events);
+    const result = treeDataProvider.getChildren() as Promise<
+      NoteProps | undefined | null
+    >;
+
+    result.then(() => {
+      this.treeView = window.createTreeView(DendronTreeViewKey.TREE_VIEW, {
+        treeDataProvider,
+        showCollapseAll: true,
       });
-      //TODO: Resolve sort order
-      // this.sortChildren(children, engine.notes);
-      return Promise.resolve(childrenNoteProps);
-    } else {
-      Logger.info({ ctx, msg: "reconstructing tree: enter" });
-      const out = Promise.all(
-        roots.flatMap(async (root) => {
-          const treeNote = await this.parseTree(root, engine.notes);
-          return treeNote.note;
-        })
+
+      this._handler = window.onDidChangeActiveTextEditor(
+        this.onOpenTextDocument,
+        this
       );
-      Logger.info({ ctx, msg: "reconstructing tree: exit" });
-      return out;
-    }
-  }
-
-  getParent(id: NoteProps): ProviderResult<NoteProps> {
-    const { engine: client } = ExtensionProvider.getDWorkspace();
-
-    const maybeParent = client.notes[id.parent || ""];
-    return maybeParent || null;
-  }
-
-  private createTreeNote(note: NoteProps) {
-    const collapsibleState = _.isEmpty(note.children)
-      ? vscode.TreeItemCollapsibleState.None
-      : vscode.TreeItemCollapsibleState.Collapsed;
-    const tn = new TreeNote({
-      note,
-      collapsibleState,
     });
-    if (note.stub) {
-      tn.iconPath = new ThemeIcon(ICONS.STUB);
-    } else if (note.schema) {
-      tn.iconPath = new ThemeIcon(ICONS.SCHEMA);
-    }
-    return tn;
   }
 
-  private async parseTree(
-    note: NoteProps,
-    ndict: NotePropsDict
-  ): Promise<TreeNote> {
-    const ctx = "parseTree";
-    const tn = this.createTreeNote(note);
-    this._tree[note.id] = tn;
-    const children = note.children;
-    this.sortChildren(children, ndict);
-    tn.children = await Promise.all(
-      children.map(async (c) => {
-        const childNote = ndict[c];
-        if (!childNote) {
-          const payload = {
-            msg: `no childNote found: ${c}, current note: ${note.id}`,
-            fullDump: _.values(ndict).map((n) => NoteUtils.toLogObj(n)),
-          };
-          const err = new DendronError({
-            message: "error updating tree view",
-            payload,
-          });
-          Logger.error({ ctx, error: err });
-          throw err;
-        }
-        return (await this.parseTree(childNote, ndict)).id;
+  /**
+   * Whenever a new note is opened, we move the tree view focus to the newly
+   * opened note.
+   * @param editor
+   * @returns
+   */
+  private async onOpenTextDocument(editor: TextEditor | undefined) {
+    if (
+      _.isUndefined(editor) ||
+      _.isUndefined(this.treeView) ||
+      !this.treeView.visible
+    ) {
+      return;
+    }
+
+    const uri = editor.document.uri;
+    const basename = path.basename(uri.fsPath);
+    const { wsRoot, vaults, engine } = ExtensionProvider.getDWorkspace();
+
+    if (
+      !WorkspaceUtils.isPathInWorkspace({
+        vaults,
+        wsRoot,
+        fpath: uri.fsPath,
       })
-    );
-    // tn.children = this.sort(children).map((c) => c.id);
-    // Logger.debug({ ctx, msg: "exit" });
-    return tn;
+    ) {
+      return;
+    }
+    if (basename.endsWith(".md")) {
+      const vault = VaultUtils.getVaultByFilePath({
+        fsPath: uri.fsPath,
+        wsRoot,
+        vaults,
+      });
+      const fname = NoteUtils.uri2Fname(uri);
+
+      const note = NoteUtils.getNoteByFnameFromEngine({
+        fname,
+        vault,
+        engine,
+      });
+
+      if (note) {
+        this.treeView.reveal(note);
+      }
+    }
   }
 }
-
-// export class TreeNoteV2 extends vscode.TreeItem {
-//   public id: string;
-//   public note: NoteProps;
-//   public uri: Uri;
-//   public children: string[] = [];
-//   public L: typeof Logger;
-
-//   constructor({
-//     note,
-//     collapsibleState,
-//   }: {
-//     note: NoteProps;
-//     collapsibleState: vscode.TreeItemCollapsibleState;
-//   }) {
-//     super(DNodeUtils.basename(note.fname, true), collapsibleState);
-//     this.note = note;
-//     this.id = this.note.id;
-//     this.tooltip = this.note.title;
-//     const vpath = vault2Path({
-//       vault: this.note.vault,
-//       wsRoot: getDWorkspace().wsRoot,
-//     });
-//     this.uri = Uri.file(path.join(vpath, this.note.fname + ".md"));
-//     if (DNodeUtils.isRoot(note)) {
-//       this.label = `root (${VaultUtils.getName(note.vault)})`;
-//     }
-//     this.command = {
-//       command: DENDRON_COMMANDS.GOTO_NOTE.key,
-//       title: "",
-//       arguments: [
-//         {
-//           qs: this.note.fname,
-//           mode: "note",
-//           vault: this.note.vault,
-//         } as GotoNoteCommandOpts,
-//       ],
-//     };
-//     this.L = Logger;
-//   }
-// }
